chore(split-panel): renames to split-pane
