+++ src/components/split-pane/split-pane.ios.scss
@@ -1,10 +1,10 @@
 
 @import "../../themes/ionic.globals.ios";
 
-// Split Panel
+// Split Pane
 // --------------------------------------------------
 
-.split-panel-ios.split-panel-visible >.split-panel-side {
+.split-pane-ios.split-pane-visible >.split-pane-side {
   min-width: 270px;
   max-width: 28%;
