+++ src/components/split-pane/split-pane.md.scss
@@ -1,10 +1,10 @@
 
 @import "../../themes/ionic.globals.md";
 
-// Split Panel
+// Split Pane
 // --------------------------------------------------
 
-.split-panel-md.split-panel-visible >.split-panel-side {
+.split-pane-md.split-pane-visible >.split-pane-side {
   min-width: 270px;
   max-width: 28%;
