+++ src/components/split-pane/split-pane.scss
@@ -1,10 +1,10 @@
 
 @import "../../themes/ionic.globals";
 
-// Split Panel
+// Split Pane
 // --------------------------------------------------
 
-ion-split-panel {
+ion-split-pane {
   position: absolute;
   top: 0;
   right: 0;
@@ -18,12 +18,12 @@ ion-split-panel {
   contain: strict;
 }
 
-.split-panel-side:not(ion-menu) {
+.split-pane-side:not(ion-menu) {
   display: none;
 }
 
-.split-panel-visible >.split-panel-side,
-.split-panel-visible >.split-panel-main {
+.split-pane-visible >.split-pane-side,
+.split-pane-visible >.split-pane-main {
   // scss-lint:disable ImportantRule
   position: relative;
   top: 0;
@@ -37,25 +37,25 @@ ion-split-panel {
   box-shadow: none !important;
 }
 
-.split-panel-visible >.split-panel-side {
+.split-pane-visible >.split-pane-side {
   flex-shrink: 0;
 
   order: -1;
 }
 
-.split-panel-visible >.split-panel-main,
-.split-panel-visible >ion-nav.split-panel-side,
-.split-panel-visible >ion-tabs.split-panel-side,
-.split-panel-visible >ion-menu.menu-enabled {
+.split-pane-visible >.split-pane-main,
+.split-pane-visible >ion-nav.split-pane-side,
+.split-pane-visible >ion-tabs.split-pane-side,
+.split-pane-visible >ion-menu.menu-enabled {
   display: block;
 }
 
-.split-panel-visible >ion-split-panel.split-panel-side,
-.split-panel-visible >ion-split-panel.split-panel-main {
+.split-pane-visible >ion-split-pane.split-pane-side,
+.split-pane-visible >ion-split-pane.split-pane-main {
   display: flex;
 }
 
-.split-panel-visible >ion-menu.menu-enabled {
+.split-pane-visible >ion-menu.menu-enabled {
   >.menu-inner {
     // scss-lint:disable ImportantRule
     right: 0;
