+++ src/components/split-pane/split-pane.wp.scss
@@ -1,10 +1,10 @@
 
 @import "../../themes/ionic.globals.wp";
 
-// Split Panel
+// Split Pane
 // --------------------------------------------------
 
-.split-panel-wp.split-panel-visible >.split-panel-side {
+.split-pane-wp.split-pane-visible >.split-pane-side {
   min-width: 270px;
   max-width: 28%;
