+++ src/components/split-pane/test/basic/e2e.ts
