+++ src/components/split-pane/test/menus/e2e.ts
