+++ src/components/split-pane/test/nested/e2e.ts
