+++ src/components/split-pane/test/nested/main.html
@@ -1,5 +1,5 @@
 
-<ion-split-panel when="sm">
+<ion-split-pane when="sm">
 
   <ion-menu [content]="content">
     <ion-nav [root]="root2"></ion-nav>
@@ -7,4 +7,4 @@
 
   <ion-nav [root]="root" main #content></ion-nav>
 
-</ion-split-panel>
+</ion-split-pane>
