+++ src/components/split-pane/test/tabs/app.module.ts
