+++ src/components/split-pane/test/tabs/e2e.ts
