+++ src/components/split-pane/test/tabs/main.html
@@ -1,8 +1,8 @@
 
-<ion-split-panel>
+<ion-split-pane>
   <ion-nav [root]="root2" #content main></ion-nav>
   <ion-tabs>
     <ion-tab [root]="root" tabTitle="Page1"></ion-tab>
     <ion-tab [root]="root" tabTitle="Page2"></ion-tab>
   </ion-tabs>
-</ion-split-panel>
+</ion-split-pane>
