+++ src/index.ts
@@ -106,7 +106,7 @@ import { ShowWhen, HideWhen } from './components/show-hide-when/show-hide-when';
 import { Slide } from './components/slides/slide';
 import { Slides } from './components/slides/slides';
 import { Spinner } from './components/spinner/spinner';
-import { SplitPanel } from './components/split-panel/split-panel';
+import { SplitPane } from './components/split-pane/split-pane';
 import { Tab } from './components/tabs/tab';
 import { Tabs } from './components/tabs/tabs';
 import { TabButton } from './components/tabs/tab-button';
@@ -192,7 +192,7 @@ export { ShowWhen, HideWhen, DisplayWhen } from './components/show-hide-when/sho
 export { Slide } from './components/slides/slide';
 export { Slides } from './components/slides/slides';
 export { Spinner } from './components/spinner/spinner';
-export { SplitPanel } from './components/split-panel/split-panel';
+export { SplitPane } from './components/split-pane/split-pane';
 export { Tab } from './components/tabs/tab';
 export { TabButton } from './components/tabs/tab-button';
 export { TabHighlight } from './components/tabs/tab-highlight';
@@ -375,7 +375,7 @@ export { Transition } from './transitions/transition';
     Slide,
     Slides,
     Spinner,
-    SplitPanel,
+    SplitPane,
     Tab,
     Tabs,
     TabButton,
@@ -470,7 +470,7 @@ export { Transition } from './transitions/transition';
     Slide,
     Slides,
     Spinner,
-    SplitPanel,
+    SplitPane,
     Tab,
     Tabs,
     TabButton,
