+++ src/themes/ionic.components.scss
@@ -191,10 +191,10 @@
 "../components/slides/slides";
 
 @import
-"../components/split-panel/split-panel",
-"../components/split-panel/split-panel.ios",
-"../components/split-panel/split-panel.md",
-"../components/split-panel/split-panel.wp";
+"../components/split-pane/split-pane",
+"../components/split-pane/split-pane.ios",
+"../components/split-pane/split-pane.md",
+"../components/split-pane/split-pane.wp";
 
 @import
 "../components/spinner/spinner",
